02-kafka-beginner-project-python/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
cramjam==2.11.0
events==0.5
idna==3.11
kafka-python==2.1.2
opensearch-dsl==2.1.0
opensearch-py==3.0.0
python-dateutil==2.9.0.post0
python-snappy==0.7.3
requests==2.32.5
six==1.17.0
sseclient-py==1.8.0
urllib3==2.5.0
websocket-client==1.6.2
